Haven't commited in a long time... :baby_chick:

+++ Ontology/ORelation.cpp
@@ -3,9 +3,9 @@
 #include "OutOfTheCave_01.h"
 #include "Ontology/ORelation.h"
 
-float ORelation::LOW_APPRECIATION = 25;
-float ORelation::REGULAR_APPRECIATION = 50;
-float ORelation::HIGH_APPRECIATION = 75;
+int ORelation::LOW_APPRECIATION = 25;
+int ORelation::REGULAR_APPRECIATION = 50;
+int ORelation::HIGH_APPRECIATION = 75;
 
 ORelation::ORelation()
 {
